Adding vpc.tf file
This file content allows the creation of a new VPC with a public and
private network, where the public has a internet gateway configuration.
The idea here is to have frontend and load balancers in the public
subnet while keeping the backend and database away from the internet
access but reachable by public services.