Adding two extra script: one for loading the postgres dump and the other for building and pushing docker image
This two script are separated in load_postgres_dump.sh and
build_push_docker_image.sh. The first will do a simple connection to the
RDS PostgreSQL address, create a rates table and load the rates.sql dump
file. The second one will build a new image, tag it and push to the ECR
repository. One important thing, the triggers here configured for the
null_resource is the current timestamp, which will make both scripts to
run every time terraform apply is called.